requirements.txt
# Dev environment
pip
autopep8
python-dotenv

# App
flask
flask-sqlalchemy
flask-login
python-telegram-bot
APScheduler
pytz
pyTelegramBotAPI
schedule
mcstatus
gunicorn
